ordanized code and updated readme,bugs
