+++ dev_data.json
@@ -1,1 +1,1 @@
-{"item":[],"editor":"vim"}
+{"item":[],"editor":"nvim"}
