+++ v1/BUGS.md
@@ -6,3 +6,4 @@
 - [x] error when list is empty
 - [ ] after changing editor it opens file in old choice even though file changed and it displays it corretly, reopening app fixes issue. 
 - [x] binary looks for file in the directory its been runed not in the place where it exist
+- [ ] when u type new item but don't confirm it dose not go away.
